# backend/docs/swagger.yaml
definitions:
...
  repo.LoginUser:
    properties:
      access_token:
        type: string
      address:
        type: string
      address_id:
        type: string
      avatar:
        type: string
      client_id:
        type: string
      email:
        type: string
      location:
        type: string
      logo:
        type: string
      name:
        type: string
      refresh_token:
        type: string
      type:
        type: string
      uid:
        type: string
    type: object
...
  server.EmailVcodeParam:
    properties:
      email:
        example: test@example.com
        type: string
    required:
    - email
    type: object
...
  server.ModifyPasswordParam:
    properties:
      new_password:
        example: Hello456
        type: string
      password:
        example: Hello123
        type: string
    required:
    - new_password
    - password
    type: object
  server.ProfileParam:
    properties:
      avatar:
        example: ""
        maxLength: 1024
        type: string
      contact_email:
        example: public@example.com
        maxLength: 256
        type: string
      description:
        example: More information about us.
        maxLength: 2048
        type: string
      github:
        example: ""
        maxLength: 128
        type: string
      location:
        example: Shanghai,China
        maxLength: 128
        type: string
      logo:
        example: ""
        maxLength: 1024
        type: string
      name:
        example: example name
        maxLength: 128
        type: string
      slack:
        example: ""
        maxLength: 128
        type: string
      twitter:
        example: ""
        maxLength: 128
        type: string
    type: object
...
  server.ResetPwdParams:
    properties:
      email:
        example: test@example.com
        type: string
      new_password:
        example: Hello456
        type: string
      vcode:
        example: "123456"
        type: string
    required:
    - email
    - new_password
    - vcode
    type: object
...
  server.SignUpParam:
    properties:
      address:
        description: 'example privateKey hex: 7b2254797065223a22736563703235366b31222c22507269766174654b6579223a22347179396744327177724f68764148305267667738624b346b52546e575535595337622f4338374435686f3d227d'
        example: f1gxcq2s72oepgufqrkbblbwgnxosrkwn3jib3bmy
        minLength: 3
        type: string
      email:
        example: test@example.com
        type: string
      message:
        description: |-
          raw data:
          Signature for filfind
          f1gxcq2s72oepgufqrkbblbwgnxosrkwn3jib3bmy
          2022-04-14T12:03:45.169Z
          raw data hex: 5369676e617475726520666f722066696c66696e640d0a663167786371327337326f657067756671726b62626c6277676e786f73726b776e336a696233626d790d0a323032322d30342d31345431323a30333a34352e3136395a
        example: 5369676e617475726520666f722066696c66696e640d0a663167786371327337326f657067756671726b62626c6277676e786f73726b776e336a696233626d790d0a323032322d30342d31345431323a30333a34352e3136395a
        type: string
      name:
        example: hello
        maxLength: 128
        type: string
      password:
        example: Hello123
        type: string
      signature:
        example: 01032aa043dea7ef185a3f5345d8bc9ee672b91584b090fca782e04f1bf36792211addd36fe6cb1e8829045d95175a48a3e431061362044bf830275829d9a7fe0d00
        type: string
      type:
        enum:
        - sp_owner
        - data_client
        example: sp_owner
        type: string
    required:
    - address
    - email
    - message
    - name
    - password
    - signature
    - type
    type: object
  server.UserParam:
    properties:
      email:
        example: test@example.com
        type: string
      password:
        example: Hello123
        type: string
    required:
    - email
    - password
    type: object
...
info:
  contact: {}
paths:
...
  /user/modifyPassword:
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer YOUR_JWT
        description: jwt access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: modify password param
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/server.ModifyPasswordParam'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: User modify password
      tags:
      - user
  /user/profile:
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer YOUR_JWT
        description: jwt access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: modify profile param
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/server.ProfileParam'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: User modify profile
      tags:
      - user
  /user/token:
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer YOUR_JWT
        description: jwt refresh token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: access token
          schema:
            type: string
      summary: Generate an access token
      tags:
      - user
  /userLogin:
    post:
      consumes:
      - application/json
      parameters:
      - description: user param
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/server.UserParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.LoginUser'
      summary: User login
      tags:
      - public
  /userResetPwd:
    post:
      consumes:
      - application/json
      parameters:
      - description: reset password param
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/server.ResetPwdParams'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: User reset password
      tags:
      - public
  /userSignUp:
    post:
      consumes:
      - application/json
      parameters:
      - description: sign up param
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/server.SignUpParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.LoginUser'
      summary: User sign up
      tags:
      - public
  /vcodeByEmailToResetPwd:
    post:
      consumes:
      - application/json
      parameters:
      - description: user email param
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/server.EmailVcodeParam'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Send a verification code by email to reset the password
      tags:
      - public
...
swagger: "2.0"